+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/oldresult/ResolvedGraphResults.java
@@ -26,5 +26,5 @@ public interface ResolvedGraphResults {
 
     Set<UnresolvedDependency> getUnresolvedDependencies();
 
-    ModuleDependency getModuleDependency(long nodeId);
+    ModuleDependency getModuleDependency(int nodeId);
 }
