+++ subprojects/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/ivyservice/DefaultLenientConfigurationTest.groovy
@@ -147,7 +147,7 @@ class DefaultLenientConfigurationTest extends Specification {
         }
 
         @Override
-        Long getNodeId() {
+        Integer getNodeId() {
             throw new UnsupportedOperationException()
         }
